+++ django_swagger_tester/loaders.py
@@ -2,7 +2,7 @@ import json
 import logging
 import os
 from json import dumps, loads
-from typing import Any, Optional, Union
+from typing import Any, Dict, Optional, Union
 
 from django.apps import apps
 from django.core.exceptions import ImproperlyConfigured
@@ -56,9 +56,7 @@ class _LoaderBase:
         """
         Sets self.schema and self.original_schema.
         """
-        self.schema = self.replace_refs(
-            schema,
-        )
+        self.schema = self.replace_refs(schema, loader=self)
         self.original_schema = schema
 
     def get_route(self, route: str) -> Route:
@@ -232,7 +230,7 @@ class _LoaderBase:
             raise ImproperlyConfigured('`status_code` should be a valid HTTP response code.')
 
     @staticmethod
-    def replace_refs(schema: dict) -> dict:
+    def replace_refs(schema: dict, loader=None) -> dict:
         """
         Finds all schema references ($ref sections) in an OpenAPI schema and inserts them back into in place of the refs.
         This way we don't have to handle reference section when interacting with a loaded schema.
@@ -240,18 +238,43 @@ class _LoaderBase:
         * This does add a performance overhead to interacting with a schema, so changing this in the future would be fine *
 
         :param schema: OpenAPI schema
+        :param loader: Current loader
         :return Adjusted OpenAPI schema
         """
         if '$ref' not in str(schema):
             return schema
+
+        schemas_cache: Dict[str, dict] = {}
 
         def find_and_replace_refs_recursively(d: dict, s: dict) -> dict:
             """
             Iterates over a dictionary to look for pesky $refs.
             """
             if '$ref' in d:
-                indices = [i for i in d['$ref'][d['$ref'].index('#') + 1 :].split('/') if i]
-                temp_schema = s
+                index_file, _, index_path = d['$ref'].partition('#')
+                # index_path looks like `/GoodTrucksList`
+                # index_file looks like `openapi-schema-split-datastructures.yaml` and will
+                # be blank unless the openapi schema is split into several sections
+                temp_schema: dict
+                if not index_file:
+                    # If index-file is empty, use local
+                    temp_schema = s
+                elif loader:
+                    # If index-file is provided, look in the cache
+                    full_path = os.path.join(os.path.dirname(loader.path), index_file)
+                    schema_from_cache: Optional[dict] = schemas_cache.get(full_path)
+                    if schema_from_cache is not None:
+                        temp_schema = schema_from_cache
+                    else:
+                        # If we cannot find our reference in the cache load external defs
+                        external_loader = loader.__class__()
+                        external_loader.set_path(full_path)
+                        temp_schema = external_loader.get_schema()
+                        schemas_cache[full_path] = temp_schema
+                else:
+                    raise RuntimeError('loader is required for external references')
+
+                indices = [i for i in index_path.split('/') if i]
                 for index in indices:
                     logger.debug('Indexing schema by `%s`', index)
                     temp_schema = temp_schema[index]