+++ tests/test_testers/test_validate_response/__init__.py
@@ -1,0 +1,37 @@
+GOOD_TEST_DATA = [
+    {
+        'url': '/cars/correct',
+        'expected_response': [
+            {'name': 'Saab', 'color': 'Yellow', 'height': 'Medium height', 'width': 'Very wide', 'length': '2 meters'},
+            {'name': 'Volvo', 'color': 'Red', 'height': 'Medium height', 'width': 'Not wide', 'length': '2 meters'},
+            {'name': 'Tesla', 'color': 'black', 'height': 'Medium height', 'width': 'Wide', 'length': '2 meters'},
+        ],
+    },
+    {
+        'url': '/trucks/correct',
+        'expected_response': [
+            {'name': 'Saab', 'color': 'Yellow', 'height': 'Medium height', 'width': 'Very wide', 'length': '2 meters'},
+            {'name': 'Volvo', 'color': 'Red', 'height': 'Medium height', 'width': 'Not wide', 'length': '2 meters'},
+            {'name': 'Tesla', 'color': 'black', 'height': 'Medium height', 'width': 'Wide', 'length': '2 meters'},
+        ],
+    },
+]
+
+BAD_TEST_DATA = [
+    {
+        'url': '/cars/incorrect',
+        'expected_response': [
+            {'name': 'Saab', 'color': 'Yellow', 'height': 'Medium height'},
+            {'name': 'Volvo', 'color': 'Red', 'width': 'Not very wide', 'length': '2 meters'},
+            {'name': 'Tesla', 'height': 'Medium height', 'width': 'Medium width', 'length': '2 meters'},
+        ],
+    },
+    {
+        'url': '/trucks/incorrect',
+        'expected_response': [
+            {'name': 'Saab', 'color': 'Yellow', 'height': 'Medium height'},
+            {'name': 'Volvo', 'color': 'Red', 'width': 'Not very wide', 'length': '2 meters'},
+            {'name': 'Tesla', 'height': 'Medium height', 'width': 'Medium width', 'length': '2 meters'},
+        ],
+    },
+]